InputMethodService: Fix system crash when selecting an inputMethod from the inputMethod menu
When the InputMethod menu is shown, if the call button is pressed and
the Inputmethod menu item is clicked at the same time, the Dialtacts
activity will be rejected by monkey if it does not belong to the package
on which the monkey is running. Due to this, the system crashes as there
is a race condition for the client to which the input method menu is
attached. Hence added a null check for the InputMethodInfo before performing the click
operation

CRs-Fixed: 183139