frameworks/base: Fixes issue of data being restarted during active UL UDP data transfer and no down link for a while.
In the case of both GSM and CDMA, the data connection is reinitialized
if it is observed that no data is received on the downlink, inspite of
active up link traffic for a while.

For GSM, the connection is reinitialized after performing a 'ping test'
but it was observed that PDP_WATCHDOG_PING_ADDRESS is set to null. For
CDMA, the radio is restarted even without the ping test.

This check is unnecessary because if the connection is really down,
then lower layers can handle this.

CRs-Fixed: 191551