Telephony: Fix StringIndexOutOfBounds Exception in CDMA mode during init.
Since the PROPERTY_OPERATOR_NUMERIC is not set until system is acquired
the value is "" when CDMAPhone is initialized. The constructor tries to
extract MCC and MNC from this property and results in a
StringIndexOutOfBounds exception. To fix this check for empty
OperatorNumeric string before trying to extract MCC and MNC and after
CDMA system is acquired, update the property and MCC, MNC configuration.