frameworks/base/telephony: Correction in EF_SPDI processing.
PLMN name is displayed in the idle screen even if the display condition in
EF_SPN file does not allow it. This is because display rule always allows
PLMN display. The contents of EF_SPDI are not processed properly. Because
of this the display rule always allows PLMN display. Corrected the EF_SPDI
data parsing.

CRs-Fixed: 220873