InputMethodManagerService: Fix acore crash on boot up
When building the input method list, if the default Input Method is not set,
the InputMethodManagerService clears the current input method and disables
the status bar input method icon.  Here, the service is trying to disable the icon
even before the statusbar variable and the input method icon are initialized.
Hence, initializing them before building the input method list.

CRs-Fixed: 223257