Telephony: Consider display condition in EF_SPN to display registered PLMN.
The display condition in EF_SPN should be considered for displaying registered
operator name even when EONS is enabled. If the first bit(b1) in the display
condition byte is set, then the registered operator name should be displayed.
Otherwise it should not be displayed.
The SIM type (whether 2G or 3G) is dynamically detected.

CRs-Fixed: 224655