Telephony: Donot send NETWORK_SELCTION Manual/Automatic to RIL on SIM READY.
As per TS 23.122 section 4.4.3.1 at switch on or recovery from lack of
coverage the UE should select the registered PLMN or equivalent PLMN (if it is
available) using all access technologies that the MS is capable of and if
necessary (in the case of recovery from lack of coverage, see clause 4.5.2)
attempts to perform a Location Registration.

Android framework sends RIL_REQUEST_SET_NETWORK_SELECTION_AUTOMATIC/
RIL_REQUEST_SET_NETWORK_SELECTION_MANUAL, on SIM READY. This happens every
time during Power On/Exit Airplane Mode/REFRESH with SIM_INIT. This is not necessary.
Modem takes care of this.

CRs-Fixed: 227179