frameworks/base: Add support for mxmf format
The framework would now recognize mxmf files as a multimedia resource.

Additionally, we have taken care to update aapt (Android Asset Packaging
tool), so that it treats mxmf as a format that does not get compressed
well (like other midi formats). AAPT is used on the host while creating
.apk files.

CRs-Fixed: 194771