Telephony: Add support for Browser Termination Event Download (Envelope command) GCF testcase 27.22.7.9.1
Added support for handling proactive command SETUP_EVENT_LIST.
On receiving the proactive command terminal response is sent
to the SIM and event is sent to the Apps. After the termination
of event (e.g.Browser Termination) envelope command is sent to
the SIM. Made it generic to support other events.

Change-Id: I38de0d772cb2cbb4b32ceaff0b27800af8c145dc
CRs-Fixed: 180239