Camera: Use the parameter ext2 of NotifyCallback to indicate whether to play shutter sound or register postview buffer.
NotifyCallback has an unused paramter ext2. Use it to indicate
whether to play shutter sound or unregister the preview buffer/register
postview buffer. We dont want to unregister the preview buffer too soon
since it will lead to blanking of screen during postview. This will
ensure that the preview buffers are unregistered as late as possible.

Change-Id: Ied65eda792f24bc6c0f459382c14cbfe167275ca