AudioRecord:Race condition between record and flinger threads causing frames to be overwritten which generates corrupted media files or files with frames overlapped generated.
On .32 kernel race condition observed during the stepserver of 9 frames. Logic
was to return error in the HAL read for 9 frames and we stepserver of 9
frames with framesize set to zero. Updated the HAL to read one frame from the
driver and loop for the number of frames requested by the flinger thread to HAL.

CRs-Fixed:235920