GPS: GpsLocationProvider is changed to accept NTP time always
The logic is changed within GpsLocationProvider to accept NTP time
irrespective of the system time. This is because NTP time is guranteed
to be accurate and system time may or not be accurate depending on
whether the phone is camped on the network or not. This results in the
GPS engine getting the best estimate of time always

CRs-Fixed : 242488

Change-Id: If661e78d23e8c537a617f646359e7d51ec83e90a