GPS:Add "address" field to GPS NI notification
In some test cases, the LCS External Address needs to be displayed.The
address is added to the notification in addition to the existing client
name and requestor id. Dependency: need a change in libloc to provide
the
 address in "extras" as Address = (the addr).Related CR for Eclair:223458
Also the GSM decoder in GpsNetInitiatedHandler.java had a bug  which
caused the last letter to be dropped in some cases.

Change-Id: I8a711f2904edde1f58963dc9f5b54989c8b50beb
CRs-Fixed: 247980